Created SubsriptionType #7 Moved the Db and ViewModels folders to Models Moved Category to Db
