--- YesterdayNews/Data/ViewModels/Class.cs
@@ -1,6 +1,0 @@
-﻿namespace YesterdayNews.Data.ViewModels
-{
-    public class Class
-    {
-    }
-}
