+++ YesterdayNews/Models/ViewModels/Class.cs
@@ -1,0 +1,6 @@
+﻿namespace YesterdayNews.Models.ViewModels
+{
+    public class Class
+    {
+    }
+}
